requirements.txt
aiohttp
aiogram
emoji
peewee
pytz
pyrogram
loguru
py_expression_eval
pydantic
pymysql
pillow
psutil
cryptography
tgcrypto
betterconf
GitPython